clean up int64_t overkill, remove most sprintf warnings

+++ src/polylib.c
@@ -126,7 +126,7 @@ WindingArea
 =============
 */
 vec_t WindingArea(winding_t *w) {
-    int64_t i;
+    int32_t i;
     vec3_t d1, d2, cross;
     vec_t total;
 
@@ -142,7 +142,7 @@ vec_t WindingArea(winding_t *w) {
 
 void WindingBounds(winding_t *w, vec3_t mins, vec3_t maxs) {
     vec_t v;
-    int64_t i, j;
+    int32_t i, j;
 
     mins[0] = mins[1] = mins[2] = BIG_BOGUS_RANGE;
     maxs[0] = maxs[1] = maxs[2] = -BIG_BOGUS_RANGE;
